Split registry and files functions in different files
