+++ src/syscheckd/syscheck.c
@@ -15,7 +15,7 @@
 #include "shared.h"
 #include "syscheck.h"
 #include "rootcheck/rootcheck.h"
-#include "db/fim_db.h"
+#include "db/fim_db_files.h"
 
 // Global variables
 syscheck_config syscheck;
